DONE: Task 1. Simplify test runner from prototype
In my prototypes, the name of the file would also be the function
that would be tested.  So a file like token-tests.php would test
the token() function and have keys like:
- $source, the first function argument
- $prefix, the second function argument
- $postfix, the third function argument
- post, (optional) post process result calling function (in this case
        token( $source, $prefix, $postfix ) before comparing it with
        expect.
- expect, the expected result
- exception, (optional) instead of expect value expect this exception

Using $arg for function argument values was great for some cases but I
ended up needing access to the arguments enough that I would end up
doing everything in the post making $arg not as useful.

One nice thing that the prototypes versions did that this version does
not is use the types function to show how expect is different from
actual.  This function has been left in because print_r, var_dump do
not indicate the type (i.e. false, null, etc.), use too much whitespace,
and so on.  Seeing these values is often revealing and therefore helpful.

types( $var ) returns a string and pv( $var ) prints using type().

Tests are still arrays but now only contain keys
- desc, which describes the test
- test, which is an anonymous function.
        return true when the test passes or false

The filename containing the test has no bearing on what function is
being tested.  The filename does not need to end -tests.php.