REDO: Task 1. Simplify test runner from prototype
The keys for tests (desc, test) are not really needed, just the use the
description (desc) as the key and then value of test (an anonymous
function) as its value.

A tradeoff is that if other keys were needed for some reason, we will
be stuck and will have to rewrite a lot tests.  YAGNI.